Move models to separate file
